+++ pos/app/components/functional/useOnlineStatus.tsx
@@ -16,9 +16,3 @@ export const useOnlineStatus = () => {
     () => true,
   );
 };
-
-// TODO(toririm): デザインがダサいので何とかする
-export const OnlineStatus = () => {
-  const online = useOnlineStatus();
-  return online ? <p>✅オンライン</p> : <p>⛔オフライン</p>;
-};
