+++ pos/app/routes/_header.cashier.tsx
